Major revamp of integrated line intensity calculation. Now doing via a fit or a direct sum.
The fit method fits 1 gaussian and 2 gaussians, compares the reduced
chi-squared values and uses the F-test to figure out what fit is
better.

The sum does an initial single gaussian fit to determine the region to
integrate over and then integrates over that region.

Errors are calculated. Upper limits are indicated.